+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Added assert result method to cv test case struct
 - Added normalisation processor to image processing pipeline, using
 ``` normalised = (source / illumination) * mean(illumination) ``` to correct for broad lighting changes
- 
+- Added region of interest isolation to the normalisation processor, hard-coded bounding box parameter and aggressive gaussian blur to remove detail outside ROI
 
 ### [Changed]
 - Updated example.jpg and example.jpeg to more accurately reflect the kind of image this system will receive